update_checksum/requirements.txt
pyDataverse==0.3.1
requests==2.25.1
psycopg2-binary==2.8.3